// types/dealer.ts
export type dealer = {
    insertedDate: string;
    _id: string;
    id: string;
    name: string;
    phone: string;
    nid: string;
    needPasswordChange: boolean;
    role: string;
    status: string;
    profileImg: string;
    isDeleted: boolean;
    createdAt: string;
    updatedAt: string;
  };