// types/product.ts
export type Product = {
  insertedDate: string,
  _id: string;
  name: string;
  bnName: string;
  category: string;
  company: string;
  dealer: string;
  packageType: string;
  quantityPerPackage: number;
  stock: number;
  price: number;
  dealerCommission: number;
  ourCommission: number;
  status: string;
  image: string;
  isDeleted: boolean;
  createdAt: string;
  updatedAt: string;
};